requirements.txt
pyyaml
yara-python
yarautil @ git+https://github.com/opensec-public/yarautil@master
